Initial committ
Initial commit (I am really struggling to get this to work).

I first attempted to add a donation via peewee but I kept getting errors such as database is locked.

I then attempted to append the new entries as a dictionary to the already selected data but the data types are different and that is not working.